refactor: Migrate yaes-slf4j and yaes-test package declarations to io.yaes
